requirements.txt
torch>=2.0.0
torchvision
torchaudio
transformers>=4.30.0
SpeechRecognition>=3.10.0
pyttsx3>=2.90
numpy>=1.24.0
python-dotenv>=1.0.0
pytest>=7.0.0
azure-cognitiveservices-speech==1.31.0
python-dotenv==1.0.0
#sentencepiece==0.1.99
datasets
pandas
scikit-learn
pyyaml
logging
matplotlib
seaborn
